requirements.txt
Flask
Flask-SocketIO
bleak
gspread
google-auth
python-dotenv
pandas
eventlet
python-engineio
python-socketio






# Flask==2.3.3
# Flask-SocketIO==5.3.4
# python-dotenv==1.0.0
# bleak==0.21.1
# gspread==5.11.3
# oauth2client==4.1.3
# requests==2.31.0























# Flask==2.3.3
# Flask-SocketIO==5.3.4
# pybluez==0.23
# gspread==5.11.4
# oauth2client==4.1.3
# python-dotenv==1.0.0
# pandas==2.0.3
# eventlet==0.33.3
# python-engineio==4.6.0
# python-socketio==5.9.0



















